chore: split i18n polyfill from production one
This will allow the production (browser API based) i18n provider to
render synchronously.
