+++ knip.ts
@@ -23,6 +23,10 @@ const config: KnipConfig = {
   ignore: [
     // Ignore React Cosmos fixtures
     'src/**/*.fixture.tsx',
+    // Ignore conditionally-compiled i18n polyfill
+    'src/common/i18n.polyfill.tsx',
+    // Ignore the locale files compiled by the i18n polyfill
+    'tests/_locales_*.js',
   ],
   ignoreDependencies: [
     // Used by our browser test and automatically detected by playwright-test.
